# Conflicts:
#	README.md
#	app/build.gradle
#	app/src/main/java/com/imcys/bilibilias/base/model/user/DownloadTaskBean.kt
#	app/src/main/java/com/imcys/bilibilias/base/utils/DialogUtils.kt
#	app/src/main/java/com/imcys/bilibilias/base/utils/DownloadQueue.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/activity/AsVideoActivity.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/activity/tool/MergeVideoActivity.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/adapter/DownloadFinishTaskAd.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/DownloadFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/HomeFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/ToolFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/AsVideoViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/FragmentHomeViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/RCMDVideoModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/AsLoginBsViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/MargeVideoViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/UserViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/factory/AsLoginBsViewModelFactory.kt
#	app/src/main/res/layout/activity_merge_video.xml
#	app/src/main/res/layout/dialog_as_login_bottomsheet.xml
#	app/src/main/res/layout/fragment_download.xml
#	app/src/main/res/layout/item_download_task_finish.xml
#	common/build.gradle
#	common/src/main/java/com/imcys/bilibilias/common/base/utils/AsVideoNumUtils.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/AppDatabase.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/dao/DownloadFinishTaskDao.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/entity/DownloadFinishTaskInfo.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/repository/DownloadFinishTaskRepository.kt


Former-commit-id: ec5c8a5ada903a52dbf569c8f8df94640771f67e

+++ app/src/main/java/com/imcys/bilibilias/home/ui/adapter/BangumiSubsectionAdapter.kt
@@ -11,7 +11,7 @@ import com.imcys.bilibilias.home.ui.model.BangumiSeasonBean
 
 class BangumiSubsectionAdapter(
     val datas: MutableList<BangumiSeasonBean.ResultBean.EpisodesBean>,
-    val defaultCid: Int,
+    val defaultCid: Long,
     val onClickMethod: (datas: BangumiSeasonBean.ResultBean.EpisodesBean, position: Int) -> Unit,
 ) : RecyclerView.Adapter<BangumiSubsectionAdapter.ViewHolder>() {
